Fix build errors.
